First actually working version of #9 Docking Layout!
