+++ Cinteros.XTB.PluginTraceViewer/Controls/TraceControl.resx
